+++ packages/react/src/reality/components/GeometryEntity.tsx
@@ -90,11 +90,11 @@ export const GeometryEntity = forwardRef<EntityRefShape, GeometryEntityProps>(
             return
           }
 
-          const materialList: SpatialMaterial[] = await Promise.all(
+          const materialList = (await Promise.all(
             materials
-              ?.map(mid => ctx.resourceRegistry.get<SpatialMaterial>(mid))
+              ?.map(mid => ctx.resourceRegistry.get(mid))
               .filter(Boolean) ?? [],
-          )
+          )) as SpatialMaterial[]
           if (gen !== mutableRef.current.rebuildGen) {
             await geometry.destroy()
             return
@@ -137,11 +137,11 @@ export const GeometryEntity = forwardRef<EntityRefShape, GeometryEntityProps>(
               createGeometry(geometryOptions),
             )
 
-            const materialList: SpatialMaterial[] = await Promise.all(
+            const materialList = (await Promise.all(
               materials
-                ?.map(id => ctx!.resourceRegistry.get<SpatialMaterial>(id))
+                ?.map(id => ctx!.resourceRegistry.get(id))
                 .filter(Boolean) ?? [],
-            )
+            )) as SpatialMaterial[]
             const modelComponent = await manager.addResource(() =>
               ctx!.session.createModelComponent({
                 mesh: geometry,